+++ front/src/App.tsx
@@ -1,41 +1,19 @@
-import { useEffect } from "react";
-import Header from "./features/header/components/Header"
-import TimerPage from "./pages/TimerPage"
-import { bgCustom } from "./utils/class"
 import { ToastContainer } from "react-toastify";
-import { useAppDispatch, useAppSelector } from "./reduxStore/hooks";
-import { checkAuthToken } from "./features/session/slices/sessionSlice";
-import { changeVisible, selectVisible, selectVisibleClass } from "./features/display/visibleSlice";
+import { Route, Routes } from "react-router";
+import { MainLayout } from "./pages/layouts/MainLayout";
+import TimerPage from "./pages/TimerPage";
 
 function App() {
-  const dispatch = useAppDispatch();
-  const visibleCalss = useAppSelector(selectVisibleClass);
-  const isVisible = useAppSelector(selectVisible);
-
-  // ページがクリックされた時に発火
-  function handleClickAnywhere() {
-    if (isVisible) return;
-    dispatch(changeVisible(true));
-  }
-
-  // マウント時に認証Check
-  useEffect(() => {
-    dispatch(checkAuthToken());
-  }, [dispatch])
-
   return (
     <>
-      <div className={`min-h-screen ${bgCustom()}`} onClick={handleClickAnywhere}>
-        <div>
-          <div className={visibleCalss}>
-            <Header/>
-          </div>
-
-          <TimerPage />
-
-          <ToastContainer autoClose={3000} draggable={false} />
-        </div>
-      </div>
+      <Routes>
+        <Route element={<MainLayout />}>
+          <Route index element={<TimerPage />} />
+        </Route>
+      </Routes>
+      
+      {/* トーストメッセージの表示 */}
+      <ToastContainer autoClose={3000} draggable={false} />
     </>
   )
 }
